dev/calendar-visualization/pyproject.toml
[tool.poetry]
name = "calendar-visualization"
version = "0.1.0"
description = ""
authors = ["Samuel Mallet  <69467005+SuperMuel@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ics = "^0.7.2"
streamlit = "^1.39.0"
streamlit-calendar = "^1.2.0"
functions = {path = "../../functions", develop = true}
python-dotenv = "^1.0.1"
langchain = "^0.3.4"
langchain-openai = "^0.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
